EnergyMatching

Documentation for EnergyMatching.

EnergyMatching.barrier — Method

Σph = barrier(Σ, Σr; Σr0=nothing, kwargs...)

Solves the energy matching problem using the barrier method.

EnergyMatching.bestricc — Method

Σphr = bestricc(Σ::QuadraticOutputStateSpace, Σr::StateSpace; kwargs...)

Solves the energy matching problem, by returning the best solution of the algebraic Riccati equation.

EnergyMatching.combined — Method

fg = combined(fgo, fgc, α)

Returns the combined objective and constraint function and its gradient for given functions fgo and fgc, for a scalar weighting factor α. The combined function is given by

f = fo + α * fc
g = go + α * gc

where fo is the objective function go its gradient and fc is the constraint function with gradient gc. For the energy matching problem fgo can be retrieved by objective and fgc by constraint.

EnergyMatching.constraint — Method

fg = constraint(Σ::QuadraticOutputStateSpace, Σr::StateSpace)

Returns the (barrier-)constraint function and its gradient for the energy matching problem, in the standard format for Optim.jl.

EnergyMatching.duplication — Method

D = duplication(n)

Returns the duplication matrix of size n x n to transform the half-vectorization to a full vectorization.

EnergyMatching.matchnrg — Method

Σrem = matchnrg(Σ::PortHamiltonianStateSpace, Σr::PortHamiltonianStateSpace; solver=:BFGS, kwargs...)

Solves the energy matching problem for the solver :BFGS (default), :COSMO or :BestRicc.

using :BFGS the problem is solved by the barrier method (see barrier).

using :COSMO the problem is solved as equivalent SDP by the SDP solver COSMO.jl (see sdp).

using :BestRicc the problem is solved by returning the best solution to the algebraic Riccati equation (see bestricc).

EnergyMatching.objective — Method

fg = objective(Σ::QuadraticOutputStateSpace, Σr::StateSpace)

Returns the objective function and its gradient for the energy matching problem, in the standard format for Optim.jl.

EnergyMatching.sdp — Function

Σphr = sdp(Σ::QuadraticOutputStateSpace, Σr::StateSpace; optimizer=COSMO.Optimizer, ε=1e-8, kwargs...)

Solves the energy matching problem using semidefinite programming.

EnergyMatching.unvech — Method

M = unvech(v)

Returns the symmetric matrix M from the half-vectorized v, i.e., the inverse of vech.

EnergyMatching.vech — Method

v = vech(M)

Returns the lower triangular part of M as a vector. Aka the half-vectorization of M. For the inverse operation see unvech.