PortHamiltonianModelReduction

Documentation for PortHamiltonianModelReduction.

PortHamiltonianModelReduction.bt — Method

Σr = bt(Σ::StateSpace, r; Lx=grampd(Σ, :o), Ly=grampd(Σ, :c)')

Reduces the state dimension of the system Σ to r using standard square root balanced truncation. The cholesky factors of the Gramians can passed as optional arguments Lx and Ly. The default values are the cholesky factors of the observability and controlability Gramians.

PortHamiltonianModelReduction.phirka — Method

Σr = phirka(Σph::PortHamiltonianStateSpace, r, num_runs; tol=1e-3, max_iter=50)

Calls phirka num_runs times and returns the best result.

PortHamiltonianModelReduction.phirka — Method

Σr = phirka(Σph::PortHamiltonianStateSpace, r; tol=1e-3, max_iter=50)

Reduces the state dimension of the port-Hamiltonian system Σph to r using the iterative rational Krylov algorithm.

PortHamiltonianModelReduction.prbt — Method

Σr = prbt(Σ, r)

Reduces the state dimension of the system Σ to r using positve real balanced truncation, which is passivity preserving. Σ can be a StateSpace or a PortHamiltonianStateSpace.