QuadraticOutputSystems

Documentation for QuadraticOutputSystems.

QuadraticOutputSystems.QuadraticOutputStateSpace — Type

QuadraticOutputStateSpace{T}

An object representing a quadratic output state space system.

dx(t)/dt = Ax(t) + Bu(t)
y(t)     = Cx(t) + M(x(t)⊗x(t))

See the function qoss for a user facing constructor.

Fields:

A::Matrix{T}

B::Matrix{T}

C::Matrix{T}

M::Matrix{T}

ControlSystemsBase.gram — Method

X = gram(Σqo::QuadraticOutputStateSpace, opt::Symbol; kwargs...)

Returns the Gramian of system Σqo. If opt is :c it returns the controllability Gramian by calling the ControlSystems.gram from ControlSystems package. If opt is :o it returns the observability Gramian by solving the Lyapunov equation

A'X + XA + MPM = 0,

where P is the controllability Gramian of Σqo.

ControlSystemsBase.grampd — Method

L = grampd(Σqo::QuadraticOutputStateSpace, opt::Symbol; kwargs...)

Returns a Cholesky factor L of the Gramian of system Σqo. If opt is :c it returns the controllability Gramian by calling the ControlSystems.grampd from ControlSystems package. If opt is :o it returns the Cholesky factor of the observability Gramian by solving the Lyapunov equation

A'X + XA + MPM = 0,

where P is the controllability Gramian of Σqo.

ControlSystemsBase.lsim — Method

result = lsim(Σqo, u, t; kwargs...)

Calculate the time response of the quadratic output state-space model Σqo::QuadraticOutputStateSpace{T} by first treating it as standard state-space model and calling ControlSystems.lsim on it, and then calculating the quadratic output y_2(t) as y_2(t) = M(x(t)⊗x(t)) and add it to the linear output.

LinearAlgebra.norm — Function

norm(Σqo, p=2; kwargs...)
h2norm(Σqo; kwargs...)

Computes the H2 norm of the quadratic output state-space model Σqo::QuadraticOutputStateSpace.

QuadraticOutputSystems.h2inner — Function

h2inner(Σ1, Σ2)

Computes the H2 inner product of the quadratic output state-space models Σ1::QuadraticOutputStateSpace and Σ2::QuadraticOutputStateSpace.

QuadraticOutputSystems.qoss — Method

Σqo = qoss(A, B, C, M)
Σqo = qoss(A, B, M)

Creates a quadratic output state-space model Σqo::QuadraticOutputStateSpace{T} with matrix element type T.