EnergyMatching

Documentation for EnergyMatching.

EnergyMatching.barrier — Method

Σph = barrier(Σ, Σr; Σr0=nothing, kwargs...)

Solves the energy matching problem using the barrier method.

EnergyMatching.bestricc — Method

Σphr = bestricc(Σ::QuadraticOutputStateSpace, Σr::StateSpace; kwargs...)

Solves the energy matching problem using the best solution of the riccati equation.

EnergyMatching.constraint — Method

fg = constraint(Σ::QuadraticOutputStateSpace, Σr::StateSpace)

Returns the (barrier-)constraint function and its gradient for the energy matching problem, in the standard format for Optim.jl.

EnergyMatching.hdss — Method

Σqo = hdss(J, R, Q, G, P, S, N)
Σqo = hdss(Σph)

Converts a PortHamiltonianStateSpace to a QuadraticOutputSystem (Hamiltonian dynamic).

EnergyMatching.matchnrg — Method

Σrem = matchnrg(Σ::PortHamiltonianStateSpace, Σr::PortHamiltonianStateSpace; solver=:BFGS, kwargs...)

Solves the energy matching problem.

EnergyMatching.objective — Method

fg = objective(Σ::QuadraticOutputStateSpace, Σr::StateSpace)

Returns the objective function and its gradient for the energy matching problem, in the standard format for Optim.jl.

EnergyMatching.sdp — Function

Σphr = sdp(Σ::QuadraticOutputStateSpace, Σr::StateSpace; optimizer=COSMO.Optimizer, ε=1e-8, kwargs...)

Solves the energy matching problem using semidefinite programming.