PortHamiltonianModelReduction

Documentation for PortHamiltonianModelReduction.

PortHamiltonianModelReduction.bt — Method

Σr = bt(Σ::StateSpace, r; Lx=grampd(Σ, :o), Ly=cholesky(gram(Σ, :c)).U)

Reduces the state dimension of the system Σ to r using standard square root balanced truncation. The cholesky factors of the Gramians can be passed as optional arguments Lx and Ly. The default values are the cholesky factors of the controllability and observability Gramians.

PortHamiltonianModelReduction.phirka — Method

Σr = phirka(Σph::PortHamiltonianStateSpace, r, num_runs; tol=1e-3, max_iter=50)

Calls phirka num_runs times and returns the best result.

PortHamiltonianModelReduction.phirka — Method

Σr = phirka(Σph::PortHamiltonianStateSpace, r; tol=1e-3, max_iter=50)

Reduces the state dimension of the port-Hamiltonian system Σph to r using the iterative rational Krylov algorithm.

S. Gugercin, R. V. Polyuga, C. Beattie, and A. van der Schaft. Structure-preserving tangential interpolation for model reduction of port-{Hamiltonian} systems. Automatica J. IFAC, 48(9):1963–1974, 2012.

PortHamiltonianModelReduction.prbt — Method

Σr = prbt(Σ::StateSpace, r)

Reduces the state dimension of the system Σ to r using positve real balanced truncation, which is passivity preserving.

U. Desai and D.Pal. A transformation approach to stochastic model reduction. IEEE Trans. Automat. Control, 29(12):1097–1100, 1984.