PortHamiltonianSystems

Documentation for PortHamiltonianSystems.

PortHamiltonianSystems.PortHamiltonianStateSpace — Type

PortHamiltonianStateSpace{T}

An object representing a port-Hamiltonian state-space system.

dx(t)/dt = (J-R)Qx(t) + (G-P)u(t)
y(t)     = (G+P)'Qx(t) + (S-N)u(t)

See the function phss for a user facing constructor.

Fields:

J::Matrix{T}

R::Matrix{T}

Q::Matrix{T}

G::Matrix{T}

P::Matrix{T}

S::Matrix{T}

N::Matrix{T}

ControlSystemsBase.gram — Method

X = gram(Σph, opt; kwargs...)

Returns the Gramian of the system Σph. If opt is :c or :o it returns the controllability or observability Gramian, respectively by calling the gram from ControlSystems package. If opt is :pr_c or :pr_o it returns the positive-real controllability or observability Gramian, respectively (see prgram).

ControlSystemsBase.grampd — Method

L = grampd(Σph, opt; kwargs...)

Returns a Cholesky factor L of the Gramian of the system Σph. If opt is :c or :o it returns the controllability or observability Gramian, respectively by calling the grampd from ControlSystems.jl package. If opt is :pr_c or :pr_o it returns a Cholesky factor L of the positive-real controllability or observability Gramian, respectively.

ControlSystemsBase.ss — Method

Σ = ss(J, R, Q, G, P, S, N)
Σ = ss(Σph)

Converts a PortHamiltonianStateSpace to a standard StateSpace.

LinearAlgebra.norm — Function

norm(Σph, p=2; kwargs...)

Converts a PortHamiltonianStateSpace to a StateSpace and calls norm on it. For more details see ControlSystems.jl package.

PortHamiltonianSystems.compose — Method

A, B, C, D = compose(J, R, Q, G, P, S, N)

Composes the port-Hamiltonian matrices according to standard state-space matrices according to

A = (J - R) * Q
B = G - P
C = (G + P)' * Q
D = S - N.

PortHamiltonianSystems.decompose — Method

J, R, Q, G, P, S, N = decompose(A, B, C, D, X)

Decomposes the standard state-space matrices to port-Hamiltonian matrices according to

Q = X
J = skew(A / X)
R = -sym(A / X)
G = 0.5 * (X \ C' + B)
P = 0.5 * (X \ C' - B)
S = sym(D)
N = skew(D).

PortHamiltonianSystems.duplication — Method

D = duplication(n)

Returns the duplication matrix of size n x n.

PortHamiltonianSystems.ispsd — Method

ispsd(M)

Returns true if M is positive semi-definite, otherwise false.

PortHamiltonianSystems.isskew — Method

isskew(M)

Returns true if M is skew-symmetric, otherwise false.

PortHamiltonianSystems.issym — Method

issym(M)

Returns true if M is symmetric, otherwise false.

PortHamiltonianSystems.kyp — Method

X = kyp(Σ; kwargs...)

Returns a solution to the KYP inequality by solving the corresponding linear matrix inequality.

PortHamiltonianSystems.kypmat — Method

W = kypmat(Σ, X)

Returns the KYP matrix of the system Σ for the given matrix X.

PortHamiltonianSystems.kypmax — Method

Xmin = kypmax(Σ; kwargs...)

Returns the maximal solution to the KYP inequality by solveing the Riccati equation for the anti-stabilizing solution.

PortHamiltonianSystems.kypmin — Method

Xmin = kypmin(Σ; kwargs...)

Returns the minimal solution to the KYP inequality by solveing the Riccati equation for the stabilizing solution.

PortHamiltonianSystems.lrcf — Method

lrcf(X, trunc_tol)

Computes an approximate low-rank Cholesky-like factorization of a symmetric positive semi-definite matrix $X$ s.t. $X = Z^T Z$ (up to a prescribed tolerance trunc_tol).

PortHamiltonianSystems.phss — Method

Σph = phss(J, R, Q, G, P, S, N)
Σph = phss(J, R, Q, G)
Σph = phss(Γ, W, Q)

Creates a port-Hamiltonian state-space model Σph::PortHamiltonianStateSpace{T} with matrix element type T.

PortHamiltonianSystems.phss — Method

Σph = phss(Σ)
Σph = phss(Σ, X)

Converts a StateSpace to a PortHamiltonianStateSpace by executing decompose. If X is not provided, the minimal solution of the KYP inequality is used.

PortHamiltonianSystems.prare — Method

prare(Σ, opt, X; kwargs...)

Evaluates the positive-real algebraic Riccati equation for system Σ and candidat solution X.

If opt is :o the positive-real controllability algebraic Riccati equation is evaluated,

A'X + XA + (C' - XB) inv(D + D') (C - B'X).

If opt is :c the positive-real observability algebraic Riccati equation is evaluated,

AX + XA' + (B - XC') inv(D + D') (B' - CX).

PortHamiltonianSystems.prgram — Method

X = prgram(Σ, opt; kwargs...)

Returns the positive-real Gramian of system Σ. If opt is :c or :o it returns the positive-real controllability or positive-real observability Gramian, respectively, by solving the corresponding positive-real algebraic Riccati equation (see prare).

PortHamiltonianSystems.prgrampd — Method

L = prgrampd(Σ, opt; kwargs...)

Returns the Cholesky factor of the positive-real Gramian of system Σ (see prgram). If opt is :c or :o it returns the positive-real controllability or positive-real observability Gramian, respectively.

In the case that the solution of the positive-real algebraic Riccati equation is not positive definite (due to numerical errors), it is projected to the set of positive semi-definite matrices calling project_psd.

PortHamiltonianSystems.project_psd — Method

Mpsd = project_psd(M; eigtol=1e-8)

Returns the nearest positive semi-definite matrix to M by setting negative eigenvalues to eigtol.

PortHamiltonianSystems.skew — Method

Mskew = skew(M)

Returns the skew-symmetric part of a matrix M.

PortHamiltonianSystems.sym — Method

Msym = sym(M)

Returns the symmetric part of a matrix M.

PortHamiltonianSystems.truncation — Method

truncation(d, L, trunc_tol) -> (dr, Lr)

Computes a rank revealing factorization for a given LDL-decomposition of $S = L * \mathrm{diag}(d) * L^T$ (up to a prescribed tolerance trunc_tol) such that $L_r * diag(d_r) * L_r^T \approx S$.

PortHamiltonianSystems.unvec — Method

M = unvec(v)

Returns the matrix M from the vectorized v, i.e., the inverse of vec.

PortHamiltonianSystems.unvech — Method

M = unvech(v)

Returns the symmetric matrix M from the half-vectorized v, i.e., the inverse of vech.

PortHamiltonianSystems.vech — Method

v = vech(M)

Returns the lower triangular part of M as a vector. Aka the half-vectorization of M. For the inverse operation see unvech.